pyproject.toml
[build-system]
requires = [
    "scikit-build-core==0.11.6",
]
build-backend = "scikit_build_core.build"

[project]
name = "sweep3d"
maintainers = [{name = "Qingnan Zhou", email = "qnzhou@gmail.com"}]
description = "Implementation of 'Lifted Surfacing of Generalized Sweep Volumes'"
readme = "README.md"
requires-python = ">=3.10"
dynamic = ["version"]
dependencies = [
    "lagrange-open>=6.39"
]

[tool.scikit-build]
install.components = ["Sweep_Python_Runtime"]
wheel.packages = ["python/sweep3d"]
cmake.build-type = "Release"
build-dir = "build-python"

[tool.scikit-build.cmake.define]
SWEEP_PYTHON_BINDING = true
CMAKE_POLICY_VERSION_MINIMUM = "3.28"

[tool.scikit-build.metadata.version]
provider = "scikit_build_core.metadata.regex"
input = "python/sweep3d/__init__.py"
